+++ pkg/hex/mix/parse.go
@@ -1,16 +1,18 @@
-package lock
+package mix
 
 import (
 	"bufio"
 	"fmt"
-	"github.com/aquasecurity/go-dep-parser/pkg/log"
 	"strings"
+	"unicode"
 
 	dio "github.com/aquasecurity/go-dep-parser/pkg/io"
+	"github.com/aquasecurity/go-dep-parser/pkg/log"
 	"github.com/aquasecurity/go-dep-parser/pkg/types"
 	"github.com/aquasecurity/go-dep-parser/pkg/utils"
 )
 
+// Parser is a parser for mix.lock
 type Parser struct{}
 
 func NewParser() types.Parser {
@@ -24,23 +26,29 @@ func (Parser) Parse(r dio.ReadSeekerAt) ([]types.Library, []types.Dependency, er
 	for scanner.Scan() {
 		lineNumber++
 		line := strings.TrimSpace(scanner.Text())
-		if strings.HasPrefix(line, "%") || strings.HasPrefix(line, "}") { // skip 1st and last lines
+		name, body, ok := strings.Cut(line, ":")
+		if !ok {
+			// skip 1st and last lines
 			continue
 		}
+		name = strings.Trim(name, `"`)
 
 		// dependency format:
 		// "<depName>": {<:hex|:git>, :<depName>, "<depVersion>", "<checksum>", [:mix], [<required deps>], hexpm", "<checksum>"},
-		ss := strings.Split(line, ", ")
+		ss := strings.FieldsFunc(body, func(r rune) bool {
+			return unicode.IsSpace(r) || r == ','
+		})
 		if len(ss) < 8 { // In the case where <required deps> array is empty: s == 8, in other cases s > 8
 			// git repository doesn't have dependency version
 			// skip these dependencies
 			if !strings.Contains(ss[0], ":git") {
-				log.Logger.Warnf("Cannot parse dependency from: %s", line)
+				log.Logger.Warnf("Cannot parse dependency: %s", line)
+			} else {
+				log.Logger.Debugf("Skip git dependencies: %s", name)
 			}
 			continue
 		}
-		name := strings.TrimLeft(ss[1], ":")
-		version := strings.Trim(ss[2], "\"")
+		version := strings.Trim(ss[2], `"`)
 		libs = append(libs, types.Library{
 			ID:        fmt.Sprintf("%s@%s", name, version),
 			Name:      name,
