+++ pkg/hex/mix/parse_test.go
@@ -1,4 +1,4 @@
-package lock
+package mix
 
 import (
 	"os"
